Pedantic buffer overflow bug fix
We shouldn’t ever run into this bug, but the default value for
`module_bounds_length` (IE, when neither PLATFORM_ID==103 or
PLATFORM_ID==269) should be 3, not 4. Leaning at 4 would create an out
of bound memory access fault — if we ever compiled that version, which
I’m sure we won’t. But still. (Pedantic!)